Prior stable release of wxWidgets _WILL_NOT_ build/compile on the ARM64-based Macs (M1). So we must upgrade to the latest stable release.
DO NOT attempt to use cmake to configure wxwidgets!!! It does not work (but
it seems like it should)!

In FileHistory.cpp::AddFileToHistory, replaced == with .Cmp for string
comparison.

main.cpp::removeFromAllWindowMenus aborts on the Mac OS Darwin arm64, but
not linux. Fortunately, it is not necessary for Mac OS. It already takes
care of this functionality for us.
I also noticed there are nested fors:
  for (int i ...
    for (int i ...
This may be problematic. (But it seems to work on Linux.)
Also, since all frames (except for MainFrame itself) are a subclass of
MainFrame, we only need to use this superclass.

Viewnix.h near line 50: changed
  #if defined (WIN32)  || defined (_WIN32)
to
  #if defined (WIN32)  || defined (_WIN32) || __MACH__